Refa. Moved Input Validation
