+++ input-validation/validateRadius.js
@@ -1,0 +1,13 @@
+exports.validateRadius = function validateRadius(radius) {
+  if (typeof radius !== "number") {
+    throw new Error(
+      `ERROR! Radius has to be a positive number but was: ${typeof radius}`
+    );
+  }
+
+  if (radius <= 0) {
+    throw new Error(
+      `ERROR! Radius has to be a positive number but was: ${radius}`
+    );
+  }
+};
